Separate table with data to TableOutput component
